+++ libminifi/include/io/tls/SecureDescriptorStream.h
@@ -61,7 +61,7 @@ class SecureDescriptorStream : public io::BaseStream {
    */
   void seek(uint64_t offset) override;
 
-  const uint64_t getSize() const override {
+  const size_t getSize() const override {
     return -1;
   }
